+++ nufftd_spread_mx.c
@@ -133,5 +133,5 @@ void mexFunction(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[])
 
     prepare_outputs(plhs, N, d, m, &tau_re, &tau_im);
 
-    nufftd_spread(tau_re, tau_im, N, n, d, omega, alpha_re, alpha_im, b, q, m);
+    nufftd_spread(tau_re, tau_im, (size_t) N, (size_t) n, (size_t) d, omega, alpha_re, alpha_im, b, q, m);
 }
