+++ sub_snufftd_spread_mx.c
@@ -188,7 +188,7 @@ void mexFunction(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[])
 
     prepare_outputs(plhs, N, d, m, &tau_re, &tau_im);
 
-    sub_snufftd_spread(tau_re, tau_im, N, n, d, grid_shift, omega, alpha_re, alpha_im, b, q, m, precomp);
+    sub_snufftd_spread(tau_re, tau_im, (size_t) N, (size_t) n, (size_t) d, grid_shift, omega, alpha_re, alpha_im, b, q, m, precomp);
 
     free_inputs(grid_shift);
 }
